+++ Mailer/extensions/mailer-module/src/Models/HealthChecker.php
@@ -3,46 +3,20 @@ declare(strict_types=1);
 
 namespace Remp\MailerModule\Models;
 
-use Predis\Client;
-
 class HealthChecker
 {
+    use RedisClientTrait;
+
     private const REDIS_KEY = 'mailer-healthcheck-';
 
-    /** @var Client */
-    private $redis;
-
-    private $host;
-
-    private $port;
-
-    private $db;
-
-    public function __construct(string $host = '127.0.0.1', int $port = 6379, int $db = 0)
+    public function __construct(RedisClientFactory $redisClientFactory)
     {
-        $this->host = $host;
-        $this->port = $port;
-        $this->db = $db;
-    }
-
-    private function connect(): Client
-    {
-        if (!$this->redis) {
-            $this->redis = new Client([
-                'scheme' => 'tcp',
-                'host'   => $this->host,
-                'port'   => $this->port,
-            ]);
-
-            $this->redis->select($this->db);
-        }
-
-        return $this->redis;
+        $this->redisClientFactory = $redisClientFactory;
     }
     
     public function ping(string $processId, int $ttlSeconds = 300): bool
     {
-        return (bool)$this->connect()->set(
+        return (bool) $this->redis()->set(
             static::REDIS_KEY . $processId,
             '1',
             'EX', // EX - Set the specified expire time, in seconds.
@@ -52,6 +26,6 @@ class HealthChecker
 
     public function isHealthy(string $processId): bool
     {
-        return $this->connect()->get(static::REDIS_KEY . $processId) !== null;
+        return $this->redis()->get(static::REDIS_KEY . $processId) !== null;
     }
 }